currency_exchage/app/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.1.7
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.4
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
